pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "nano-vllm-triton"
version = "0.1.0"
authors = [{ name = "Yu Liu" }]
license = "MIT"
license-files = ["LICENSE"]
readme = "README.md"
description = "A highly optimized lightweight vLLM implementation with custom Triton kernels"
requires-python = ">=3.10,<3.13"
dependencies = [
    "torch>=2.4.0",
    "triton>=3.0.0",
    "transformers>=4.51.0",
    "flash-attn",
    "xxhash",
    "pyyaml"
]

[project.urls]
Homepage="https://github.com/cosmoliu2002/nano-vllm-triton"

[tool.setuptools.packages.find]
where = ["."]
include = ["nanovllm*"]
